Fix for small display and operations showing up
The second number on the small display was not updating correctly so
I fixed that by adding the newest number button pressed onto the small
display. The calculator also allowed more than one operator to be
shown on the display but now it should only show one.

Changes to be committed:
	modified:   index.js