Fix calculator display after dividing by 0
After dividing by 0 and displaying "N/A", the next number buttin click
would cause the screen to go blank. If a number button was pushed for
a second time, then the number pushed would show on the display. This
fixes it to where the next number button clicked while "N/A" is on the
display shows up in place of "N/A".

 Changes to be committed:
	modified:   index.js